+++ packages/laravel/config/numera.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'default_locale' => 'en',
+    'fallback_locale' => 'en',
+];